requirements.txt
azure-cognitiveservices-speech==1.34.0
numpy==1.24.3
python-dotenv==1.0.0
sounddevice==0.4.6
scipy==1.11.1 
python-dotenv==1.0.0 